# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ChaosGuard Automated Chaos Engineering Platform API",
        "title": "ChaosGuard API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/containers": {
            "get": {
                "description": "Returns a list of target containers and their monitoring status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Containers"
                ],
                "summary": "List Containers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/responses.ContainerResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/containers/{id}": {
            "get": {
                "description": "Returns details for a single target container",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Containers"
                ],
                "summary": "Get Container",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Container ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.ContainerResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/experiments": {
            "get": {
                "description": "Returns a list of chaos experiments executed or scheduled",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experiments"
                ],
                "summary": "List Experiments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/responses.ExperimentResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Starts a manual/ad-hoc chaos attack against a container",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experiments"
                ],
                "summary": "Inject Failure",
                "parameters": [
                    {
                        "description": "Attack Details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateExperimentRequest"
                        }
                    }
                ],
                "responses": {
                    "211": {
                        "description": "",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.ExperimentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/experiments/{id}": {
            "get": {
                "description": "Returns details for a single chaos experiment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experiments"
                ],
                "summary": "Get Experiment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Experiment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.ExperimentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Stops a running experiment early and deletes its history record",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Experiments"
                ],
                "summary": "Cancel Experiment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Experiment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns the application health status and runtime state",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Get Health status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.HealthResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/runtime": {
            "get": {
                "description": "Returns the lifecycle state of the ChaosGuard daemon",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Runtime"
                ],
                "summary": "Get Runtime State",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.RuntimeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/scheduler/start": {
            "post": {
                "description": "Resumes the background chaos failure injection loop",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scheduler"
                ],
                "summary": "Start Scheduler",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/scheduler/status": {
            "get": {
                "description": "Returns the status and scheduling settings of the automated scheduler",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scheduler"
                ],
                "summary": "Get Scheduler Status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.SchedulerStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/scheduler/stop": {
            "post": {
                "description": "Pauses the automated scheduler loop",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scheduler"
                ],
                "summary": "Stop Scheduler",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "requests.CreateExperimentRequest": {
            "type": "object",
            "required": [
                "attack_type",
                "duration",
                "target_container_id"
            ],
            "properties": {
                "attack_type": {
                    "type": "string",
                    "example": "pause"
                },
                "duration": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 10
                },
                "target_container_id": {
                    "type": "string",
                    "example": "c1"
                }
            }
        },
        "responses.ContainerResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-07-12T16:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "c1"
                },
                "image": {
                    "type": "string",
                    "example": "nginx:latest"
                },
                "is_monitored": {
                    "type": "boolean",
                    "example": true
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "web"
                },
                "status": {
                    "type": "string",
                    "example": "running"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-07-12T16:05:00Z"
                },
                "uptime": {
                    "type": "string",
                    "example": "2h45m10s"
                }
            }
        },
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "responses.ExperimentResponse": {
            "type": "object",
            "properties": {
                "attack_type": {
                    "type": "string",
                    "example": "pause"
                },
                "container_name": {
                    "type": "string",
                    "example": "web"
                },
                "duration": {
                    "type": "integer",
                    "example": 10
                },
                "ended_at": {
                    "type": "string",
                    "example": "2026-07-12T16:10:10Z"
                },
                "error_message": {
                    "type": "string",
                    "example": "attack failed"
                },
                "id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "parameters": {
                    "type": "string",
                    "example": "{\"duration\":\"10s\"}"
                },
                "recovery_status": {
                    "type": "string",
                    "example": "recovered"
                },
                "started_at": {
                    "type": "string",
                    "example": "2026-07-12T16:10:00Z"
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "target_container_id": {
                    "type": "string",
                    "example": "c1"
                }
            }
        },
        "responses.HealthResponse": {
            "type": "object",
            "properties": {
                "state": {
                    "type": "string",
                    "example": "running"
                },
                "status": {
                    "type": "string",
                    "example": "healthy"
                },
                "version": {
                    "type": "string",
                    "example": "v0.1.0-dev"
                }
            }
        },
        "responses.RuntimeResponse": {
            "type": "object",
            "properties": {
                "state": {
                    "type": "string",
                    "example": "running"
                }
            }
        },
        "responses.SchedulerStatusResponse": {
            "type": "object",
            "properties": {
                "attack_duration": {
                    "type": "string",
                    "example": "10s"
                },
                "attack_interval": {
                    "type": "string",
                    "example": "30s"
                },
                "mode": {
                    "type": "string",
                    "example": "random"
                },
                "running": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "responses.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    }
}